+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/flow/LMap.java
@@ -198,6 +198,11 @@ public class LMap extends Component implements HasSize, HasStyle
 		this.center = start;
 		this.setViewPoint(start);
 	}
+	
+	@ClientCallable
+	protected void onMapCenterChanged(final double lat, final double lng, final int zoom) {
+		this.center = new LCenter(lat, lng, zoom);
+	}
 
 
 	@ClientCallable